requirement.txt
# Discord Bot Dashboard — Requirements
# Installa con: pip install -r requirements.txt

discord.py[voice]>=2.3.0
yt-dlp>=2024.1.0
psutil>=5.9.0
PyNaCl>=1.5.0